refactor: :recycle: refactor functions
